+++ experiments/mobx_hooks/lib/main.dart
@@ -1,7 +1,0 @@
-import 'package:jaspr/server.dart';
-import './app.dart';
-import 'mobx_hooks/jaspr_observer.dart';
-
-void main() {
-  runApp(const MobXHooksObserverComponent(child: App()));
-}