+++ experiments/mobx_hooks/web/main.dart
@@ -1,7 +1,0 @@
-import 'package:jaspr/browser.dart';
-import 'package:mobx_hooks_experiment/app.dart';
-import 'package:mobx_hooks_experiment/mobx_hooks/jaspr_observer.dart';
-
-void main() {
-  runApp(const MobXHooksObserverComponent(child: App()));
-}