+++ experiments/preload_images/lib/services/service.dart
@@ -1,9 +1,0 @@
-// example service
-
-abstract class ImageService {
-  // better to use a service locator, this is just for demo purposes
-  static ImageService? instance;
-
-  Future<Map<String, dynamic>> getImageById(String id);
-  Future<List> getImages();
-}