+++ experiments/scoped_styles/web/main.dart
@@ -1,6 +1,0 @@
-import 'package:jaspr/browser.dart';
-import 'package:scoped_styles_experiment/app.dart';
-
-void main() {
-  runApp(App());
-}