+++ packages/jaspr_lints/lib/src/assists/css_assist.dart
@@ -1,7 +1,6 @@
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer/source/source_range.dart';
-import 'package:analyzer_plugin/protocol/protocol_common.dart';
 import 'package:analyzer_plugin/utilities/change_builder/change_builder_dart.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
@@ -101,21 +100,14 @@ class CssAssistProvider extends DartAssist {
       void writeCssRule(DartEditBuilder edit) {
         edit.write('  css(\'');
         if (idVal != null) {
-          edit.write('#$idVal\').');
+          edit.write('#$idVal');
         } else if (classesVal != null) {
-          edit.write('.$classesVal\').');
+          edit.write('.$classesVal');
         } else {
           edit.write('.');
           edit.addSimpleLinkedEdit('className', 'classname');
-          edit.write('\').');
         }
-        edit.addSimpleLinkedEdit(
-          'styles',
-          'box',
-          kind: LinkedEditSuggestionKind.METHOD,
-          suggestions: ['box', 'text', 'background', 'flexbox', 'flexItem', 'grid', 'gridItem', 'list'],
-        );
-        edit.write('(),\n  ');
+        edit.write('\').styles(),\n  ');
       }
 
       if (styles == null) {