+++ packages/jaspr_riverpod/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.19
+
+- `jaspr` upgraded to `0.18.0`
+
 ## 0.3.18
 
 - Update logo and website links.