+++ packages/jaspr_router/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.4
+
+- Fixed bug with router redirects.
+
+## 0.6.3
+
+- `jaspr` upgraded to `0.18.0`
+
 ## 0.6.2
 
 - Update logo and website links.